added card types
Each type holds and enharits to eachouter so if you create a creature goblen rader it will apply vars from the super types it inherits from so for this example for creature goblin ranger the goblen and ranger are sub types so they dont add vars but they do allow for detection by other cards. What is important is the super type witch is creature, since it is a creature i inherits two vars power and health and somining sickness from it being a creature that is does have sommining sickness and it adds the invisabull tags to the creature wich are in total are Creature, Goblin, Permanent, and Spell this is because creature inherits from permanent and permanent inherits from spell. These tags are important because they alow the detection of what cards exist or are being played and for the expedient proses of creating cards.